requirements.txt
flask>=3.0.0
flask-socketio>=5.3.0
google-genai>=1.0.0
elevenlabs>=1.0.0
python-dotenv>=1.0.0
easyocr>=1.7.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
mss>=9.0.0
